Add accurate time reporting
The hit times weren't being reported correctly. Previously we would send the GA hits all at once, with each scan of the logs (which happens every 10 minutes).
The result was that all hits were being reported on the 10 minute mark, instead of as they were happening. Luckily the logs also provide a timestamp of when the event actually occurred so we can pass that along to GA.

We're using the Queue Time functionality (param `qt`) to accurately report when the actual hit happened: https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters#qt

So what we're sending in the `qt` param is the time difference between now and when the hit actually occured (in milliseconds): now - log_timestamp. This will let GA know how much time in the past it should place the hit.